Complete A-MoreOOP
Extract Ui, Storage, Parser, TaskList from Duke
Add Javadoc for the 4 extracted classes
Update Javadoc for Task and its children classes
